test(server): split reference wire tests into reference_writes.rs + indexed-read pins (#182)
- multi_element_writes.rs crossed the 700-LOC cap; the seven reference-
  property wire tests move to a self-contained reference_writes.rs (same
  convention as framed_properties.rs: local write/read/assert helpers),
  and multi_element_writes keeps the full-consumption matrix + datetime
  unlocks.
- New pins: the empty-frame and Null clears of SETPOINT_REFERENCE over
  WriteProperty (both paths, set-then-clear each); Averaging
  Object_Property_Reference over the wire (framed local write lands with
  historical Unsigned read-back; device-qualified [3] and 4-member flat
  writes refused with state preserved).
- array_index_gating.rs: the five reference-typed properties (Loop CVR /
  MVR / SR, Pulse Converter INPUT_REFERENCE, Averaging OPR) reject an
  indexed ReadProperty with PROPERTY / PROPERTY_IS_NOT_AN_ARRAY — closes
  the loop between the #182 read arms and the tranche-K gate.
